pyproject.toml
[tool.poetry]
name = "mav-utils"
version = "0.1.0"
description = ""
authors = ["Malins062 <6328673@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pyinstaller = "^6.16.0"
loguru = "^0.7.3"
pillow = "^12.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.5.2"
pyright = "^1.1.371"
setuptools = "^80.3.1"

[tool.black]
line-length = 120

[tool.ruff]
show-fixes = true
fix = true
lint.select = ["F", "E", "W", "C90", "I", "N", "S", "B", "A", "ISC", "T20", "Q", "PTH"]
lint.ignore = ["A003", "E501", "ISC001"]

[tool.pyright]
reportUnnecessaryTypeIgnoreComment="warning"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
